home fragment code cleanup,  shifting notification code to separate file
