+++ app/src/main/java/com/example/taskplanner/fragments/HomeFragment.kt
@@ -76,81 +76,6 @@ class HomeFragment : Fragment() {
             // Alert Box
             dialog()
         }
-        //findNavController().navigate(R.id.action_homeFragment_to_projectFragment)
-
-
-
-        // Alarm manager for notification
-        var taskOne : Task = Task(SystemClock.elapsedRealtime() + 60 * 1000, SystemClock.elapsedRealtime() + 120 * 1000 , "TaskOne" , "DescTaskOne" , 0)
-//        var taskTwo : Task = Task(4484884, 5998595 , "TaskOne" , "DescTaskOne" , 44484)
-//        var taskThree : Task = Task(4484884, 5998595 , "TaskOne" , "DescTaskOne" , 44484)
-
-        // List of tasks
-        val taskList = ArrayList<Task>()
-        taskList.add(taskOne)
-
-
-        // Loop through the task list and schedule notifications for each task
-        val alarmManager1 = contextApp?.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-        for (task in taskList) {
-            // Create an intent for the start time notification with a unique request code
-            val startTimeIntent = Intent(contextApp, StartTimeReceiver::class.java)
-            val startTimePendingIntent = PendingIntent.getBroadcast(
-                contextApp, 0, startTimeIntent, PendingIntent.FLAG_IMMUTABLE
-            )
-
-            // Create an intent for the end time notification with a unique request code
-            val endTimeIntent = Intent(contextApp, EndTimeReceiver::class.java)
-            val endTimePendingIntent = PendingIntent.getBroadcast(
-                contextApp, 123, endTimeIntent, PendingIntent.FLAG_IMMUTABLE
-            )
-
-            // Schedule the start time notification
-            alarmManager1?.setExactAndAllowWhileIdle(
-                AlarmManager.ELAPSED_REALTIME_WAKEUP,
-                task.startTime,
-                startTimePendingIntent
-            )
-
-            // Schedule the end time notification
-            alarmManager1?.setExactAndAllowWhileIdle(
-                AlarmManager.ELAPSED_REALTIME_WAKEUP,
-                task.endTime,
-                endTimePendingIntent
-            )
-        }
-
-
-
-
-
-
-
-
-
-//        val alarmManager = contextApp?.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-//
-//        val intent =  Intent(contextApp, StartTimeReceiver::class.java)
-//        Log.i("HomeFragment" , "Setting the pending intent")
-//        val pendingIntent : PendingIntent = PendingIntent.getBroadcast(
-//            requireContext(), 0, intent,
-//            PendingIntent.FLAG_IMMUTABLE
-//        )
-//
-//
-//        alarmManager.setExactAndAllowWhileIdle(
-//            AlarmManager.ELAPSED_REALTIME_WAKEUP,
-//            SystemClock.elapsedRealtime() + 60 * 1000,
-//            pendingIntent
-//        )
-
-
-
-
-
-
-
-
     }
 
     private fun dialog(){
@@ -235,63 +160,5 @@ class HomeFragment : Fragment() {
 
     }
 
-}
-
-class StartTimeReceiver : BroadcastReceiver() {
-    override fun onReceive(context: Context, intent: Intent) {
-        // Show notification for start time
-        var channelId = "CHANNEL_ID"
-
-        val notificationBuilder = NotificationCompat.Builder(context, channelId)
-            .setContentTitle("Task Started")
-            .setContentText("Your task has started now")
-            .setSmallIcon(R.drawable.icon_task_done)
-            .setPriority(NotificationCompat.PRIORITY_DEFAULT)
-
-        Log.i("HomeFragment" , "This is the notification Builder class")
-
-
-        // Show the notification
-        val notificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val channel = NotificationChannel(
-                channelId,
-                "Channel",
-                NotificationManager.IMPORTANCE_DEFAULT
-            )
-            notificationManager.createNotificationChannel(channel)
-        }
-        notificationManager.notify(0, notificationBuilder.build())
-
-    }
-}
-
-
-
-class EndTimeReceiver : BroadcastReceiver() {
-    override fun onReceive(context: Context, intent: Intent) {
-        // Show notification for end time
-
-        var channelId = "CHANNEL_ID"
-        val notificationBuilder = NotificationCompat.Builder(context, channelId)
-            .setContentTitle("Task Ended")
-            .setContentText("Did you finished your task")
-            .setSmallIcon(R.drawable.icon_not_finished)
-            .setPriority(NotificationCompat.PRIORITY_DEFAULT)
-
-
-// Show the notification
-        val notificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val channel = NotificationChannel(
-                channelId,
-                "Channel",
-                NotificationManager.IMPORTANCE_DEFAULT
-            )
-            notificationManager.createNotificationChannel(channel)
-        }
-        notificationManager.notify(123, notificationBuilder.build())
-
-    }
 }
